Add rabit mq into wpf application for fibonacci service

+++ FibonacciService/Shared/Core/Models/FibRequest.cs
@@ -7,6 +7,6 @@ namespace Core.Models
     {
         [Required]
         [Range(2, 100)]
-        public int NumberToCalculate { get; set; }
+        public long NumberToCalculate { get; set; }
     }
 }
