+++ FibonacciService/Shared/Core/Models/FibResultResponse.cs
@@ -9,7 +9,7 @@ namespace Core.Models
     {
         public int Id { get; set; }
 
-        public int ElementNumber { get; set; }
+        public long ElementNumber { get; set; }
 
         public long Result { get; set; }
     }
